Changed health and damage values to be used as ints

+++ Tower Defense/Assets/Scripts/AI/Interfaces/IDamageable.cs
@@ -5,5 +5,5 @@
 /// </summary>
 public interface IDamageable
 {
-    void GetDamage(float damage);
+    void GetDamage(int damage);
 }
